implementa ajustes na view e serviço 'FormataData'

+++ src/migrations/20210510231000-change-column-anuncios.js
@@ -2,14 +2,14 @@ module.exports = {
     up: (queryInterface, Sequelize) => {
         return Promise.all([
             queryInterface.changeColumn('Anuncios', 'investimentoDia', {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT,
             }),
         ])
     },
     down: (queryInterface, Sequelize) => {
         return Promise.all([
             queryInterface.changeColumn('Anuncios', 'investimentoDia', {
-                type: Sequelize.FLOAT,
+                type: Sequelize.INTEGER,
             }),
         ])
     }
